+++ core/src/main/java/org/vertexarmy/dsr/game/Level.java
@@ -2,13 +2,14 @@ package org.vertexarmy.dsr.game;
 
 import com.beust.jcommander.internal.Lists;
 import com.google.common.collect.ImmutableList;
-import java.io.Serializable;
-import java.util.List;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import org.vertexarmy.dsr.math.Polygon;
+
+import java.io.Serializable;
+import java.util.List;
 
 /**
  * created by Alex
@@ -17,6 +18,21 @@ import org.vertexarmy.dsr.math.Polygon;
 @EqualsAndHashCode
 @RequiredArgsConstructor
 public class Level implements Serializable {
+
+    public enum BackgroundLayerType {
+        BACKGROUND(0),
+        FAR(1),
+        MIDDLE(2),
+        NEAR(3),
+        FOREGROUND(4);
+
+        @Getter
+        private final int zOrder;
+
+        private BackgroundLayerType(int zOrder) {
+            this.zOrder = zOrder;
+        }
+    }
 
     public static class BackgroundLayer implements Serializable {
         @Getter
@@ -29,16 +45,16 @@ public class Level implements Serializable {
 
         @Getter
         @Setter
-        private int zOrder;
+        private BackgroundLayerType type;
 
         public BackgroundLayer(String textureName, float parallaxSpeedScale) {
-            this(textureName, parallaxSpeedScale, 0);
+            this(textureName, parallaxSpeedScale, BackgroundLayerType.BACKGROUND);
         }
 
-        public BackgroundLayer(String textureName, float parallaxSpeedScale, int zOrder) {
+        public BackgroundLayer(String textureName, float parallaxSpeedScale, BackgroundLayerType type) {
             this.textureName = textureName;
             this.parallaxSpeedScale = parallaxSpeedScale;
-            this.zOrder = zOrder;
+            this.type = type;
         }
     }
 
@@ -53,6 +69,16 @@ public class Level implements Serializable {
 
     @Getter
     private final List<Polygon> terrainPatches;
+
+    public BackgroundLayer getBackgroundLayerByType(BackgroundLayerType type) {
+        for (BackgroundLayer layer : backgroundLayers) {
+            if (layer.type == type) {
+                return layer;
+            }
+        }
+
+        return null;
+    }
 
     public static Level createDefaultLevel() {
         return new Level(null, null, ImmutableList.of(new Polygon(new float[]{100, 100, 100, -100, -100, -100, -100, 100})));